Make Fixer responsible for realigning tab stops
